requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
cryptography==3.4.8
Django==3.1.1
django-appconf==1.0.4
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-cryptographic-fields==0.5.1
django-cryptography==1.0
django-editorjs==0.2.1
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
idna==2.10
Pillow==8.1.0
pycparser==2.20
PyJWT==1.7.1
pythonpy==0.4.11
pytube==9.6.0
pytube3==9.6.4
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.25.10
